added constants

+++ src/consts.d
@@ -40,8 +40,8 @@ shared float[3][6][2] Weights = [
 
 /// Maps coefficient location to a weight bucket
 /// Initialized with a self executing lambda, a-la Javascript
-shared float[ImageArea] WeightBins = (() {
-	float[ImageArea] tmp;
+shared byte[ImageArea] WeightBins = (() {
+	byte[ImageArea] tmp;
 	foreach(i; 0..ImageHeight) {
 		foreach(j; 0..ImageWidth) {
 			tmp[(i * ImageHeight) + j] = min(max(i, j), 5);
